requirements.txt
# Core ML
numpy>=1.26.0,<3
pandas>=2.1.0,<3
scikit-learn>=1.3.0,<2
torch>=2.2.0,<3

# ANN
faiss-cpu>=1.8.0

# Serving
fastapi>=0.111.0,<1
uvicorn[standard]>=0.30.0,<1

# Utilities / testing
tqdm>=4.66.0,<5
joblib>=1.3.0,<2
pytest>=8.0.0,<9
